requirements.txt
django
djangorestframework
django-environ
gunicorn
psycopg2
